requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dropbox==12.0.2
idna==3.7
ply==3.11
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
stone==3.3.1
urllib3==2.2.2
